rustc_ast_lowering
Trait ResolverAstLoweringExt

pub(crate) trait ResolverAstLoweringExt {
    // Required methods
    fn legacy_const_generic_args(&self, expr: &Expr) -> Option<Vec<usize>>;
    fn get_partial_res(&self, id: NodeId) -> Option<PartialRes>;
    fn get_import_res(&self, id: NodeId) -> PerNS<Option<Res<NodeId>>>;
    fn get_label_res(&self, id: NodeId) -> Option<NodeId>;
    fn get_lifetime_res(&self, id: NodeId) -> Option<LifetimeRes>;
    fn extra_lifetime_params(
        &mut self,
        id: NodeId,
    ) -> Vec<(Ident, NodeId, LifetimeRes)>;
}

Required Methods

fn legacy_const_generic_args(&self, expr: &Expr) -> Option<Vec<usize>>

fn get_partial_res(&self, id: NodeId) -> Option<PartialRes>

Obtains resolution for a NodeId with a single resolution.

fn get_import_res(&self, id: NodeId) -> PerNS<Option<Res<NodeId>>>

Obtains per-namespace resolutions for use statement with the given NodeId.

fn get_label_res(&self, id: NodeId) -> Option<NodeId>

Obtains resolution for a label with the given NodeId.

fn get_lifetime_res(&self, id: NodeId) -> Option<LifetimeRes>

Obtains resolution for a lifetime with the given NodeId.

fn extra_lifetime_params( &mut self, id: NodeId, ) -> Vec<(Ident, NodeId, LifetimeRes)>

Obtain the list of lifetimes parameters to add to an item.

Extra lifetime parameters should only be added in places that can appear as a binder in LifetimeRes.

The extra lifetimes that appear from the parenthesized Fn-trait desugaring should appear at the enclosing PolyTraitRef.